+++ packages/webpack/__e2e-fixtures__/babel-postcss-extract-text/webpack.config.js
@@ -22,9 +22,11 @@ module.exports = createConfig([
   ),
   babel(),
   match('*.css', [
-    postcss([
-      precss
-    ]),
+    postcss({
+      plugins: [
+        precss
+      ]
+    }),
     extractText(
       'styles.css'
     )